refactor: AuthorDto 위치 이동
